Intermediate commit switching computers
