+++ roslibrust/src/ros1/mod.rs
@@ -3,5 +3,8 @@
 mod master_client;
 pub use master_client::*;
 
+mod node_server;
+pub(crate) use node_server::*;
+
 mod node;
 pub use node::*;
